perf: php ns segment interning
unable to use rust lifetime system because it doesn't like it if i mark
the HashSet mutable even though I only insert into it (and never
delete). i guess we have to deal with slightly more overhead now.

credits to chatgpt for misleading me into thinking this was possible
with the lifetime system.